requirements.txt
selenium-wire>=5.1.0
selenium>=4.0.0
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
